refactor: separete pokeapi and scraper in different packages
