--- api/client.go
@@ -1,5 +1,0 @@
-package api
-
-type PokeClient interface {
-	FetchPokemons(characteristic, value string) ([]byte, error)
-}
